requirements.txt
alembic==1.13.1
algoliasearch==2.6.2
algoliasearch-django==2.0.0
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.12.0
apturl==0.5.2
arabic-reshaper==3.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.2.3
asgiref==3.10.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==22.2.0
autocommand==2.2.2
backports.tarfile==1.2.0
bcrypt==3.2.0
billiard==4.2.2
binaryornot==0.4.4
blessed==1.19.1
blinker==1.4
boto==2.49.0
boto3==1.26.65
botocore==1.29.65
Brlapi==0.8.3
Brotli==1.1.0
cachetools==6.2.0
celery==5.5.3
certifi==2022.12.7
cffi==2.0.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2025.6
crispy-tailwind==1.0.3
cron_descriptor==2.0.6
cryptography==46.0.3
cssselect2==0.8.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.12.1
django-appconf==1.1.0
django-celery-beat==2.8.1
django-ckeditor==6.4.2
django-compressor==4.5.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-extensions==4.1
django-filter==22.1
django-formtools==2.5.1
django-import-export==3.1.0
django-js-asset==2.2.0
django-model-utils==5.0.0
django-phonenumber-field==8.1.0
django-picklefield==3.1
django-q-email==5.0.1
django-q2==1.8.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-rest-registration==0.7.3
django-storages==1.13.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
drf-yasg==1.21.5
duplicity==0.8.21
ecdsa==0.19.1
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.3.1
fastapi==0.110.0
fasteners==0.14.1
fido2==2.0.0
fonttools==4.60.1
future==0.18.2
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gpg==1.16.0
greenlet==3.3.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib_metadata==8.0.0
inflection==0.5.1
iniconfig==2.3.0
itypes==1.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jaraco.text==4.0.0
jeepney==0.7.1
Jinja2==3.1.2
jinxed==1.2.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.5.0
kombu==5.5.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==6.0.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==4.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
mdurl==0.1.2
mercurial==6.1.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.1
oscrypto==1.3.0
packaging==24.2
pandas==2.3.3
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
phonenumbers==9.0.14
pi==0.1.2
pillow==12.1.0
pip-chill==1.0.3
platformdirs==4.2.2
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.11
psycopg-c==3.2.11
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.23
pycryptodome==3.23.0
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pydyf==0.11.0
PyGithub==2.8.1
Pygments==2.19.2
PyGObject==3.42.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==5.9.0
pyphen==0.17.2
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==8.1.1
python-apt==2.4.0+ubuntu4
python-bidi==0.6.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2022.6
pyxdg==0.27
PyYAML==6.0.3
rave-python==1.4.1
rcssmin==1.1.2
redis==6.4.0
reportlab==3.6.8
requests==2.31.0
rich==14.2.0
rjsmin==1.2.2
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.14
s3transfer==0.6.0
SecretStorage==3.3.1
sib-api-v3-sdk==7.6.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.3
starlette==0.36.3
svglib==1.5.1
swagger-spec-validator==3.0.3
systemd-python==234
tablib==3.3.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.0.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==1.26.13
usb-creator==0.3.7
uv==0.9.5
uvicorn==0.29.0
vboxapi==1.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.5
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
zopfli==0.2.3.post1
